changing User model

+++ src/main/java/io/github/raphaelrighetti/blogsphere/controllers/UserController.java
@@ -14,6 +14,8 @@ import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import io.github.raphaelrighetti.blogsphere.models.dto.UserChangePasswordDTO;
+import io.github.raphaelrighetti.blogsphere.models.dto.UserListingDTO;
 import io.github.raphaelrighetti.blogsphere.models.dto.UserReadDTO;
 import io.github.raphaelrighetti.blogsphere.models.dto.UserUpdateDTO;
 import io.github.raphaelrighetti.blogsphere.services.CheckOwnerService;
@@ -32,35 +34,46 @@ public class UserController {
 	private CheckOwnerService checkOwnerService;
 	
 	@GetMapping
-	public ResponseEntity<Page<UserReadDTO>> get(@PageableDefault Pageable pageable) {
-		Page<UserReadDTO> page = userService.get(pageable);
+	public ResponseEntity<Page<UserListingDTO>> get(@PageableDefault Pageable pageable) {
+		Page<UserListingDTO> page = userService.get(pageable);
 		
 		return ResponseEntity.ok(page);
 	}
 	
 	@GetMapping("/{id}")
-	public ResponseEntity<UserReadDTO> getById(@PathVariable Long id) {
+	public ResponseEntity<UserReadDTO> getById(@PathVariable Long id, @RequestHeader("Authorization") String header) {
+		checkOwnerService.checkOwner(id, header);
 		UserReadDTO responseDTO = userService.getById(id);
 		
 		return ResponseEntity.ok(responseDTO);
 	}
 	
-	@PutMapping("{id}")
+	@PutMapping("/{id}")
 	@Transactional
 	public ResponseEntity<Void> update(@PathVariable Long id, @RequestBody @Valid UserUpdateDTO dto,
 			@RequestHeader("Authorization") String header) {
-		checkOwnerService.isOwner(id, header);
-		
+		checkOwnerService.checkOwner(id, header);
 		userService.update(id, dto);
 		
 		return ResponseEntity.noContent().build();
 	}
 	
-	@DeleteMapping("{id}")
+	@PutMapping("/{id}/change-password")
+	@Transactional
+	public ResponseEntity<Void> changePassword(@PathVariable Long id, @RequestBody @Valid UserChangePasswordDTO dto, 
+			@RequestHeader("Authorization") String header) {
+		checkOwnerService.checkOwner(id, header);
+		userService.changePassword(id, dto);
+		
+		// TODO: Implement verification by email logic
+		
+		return ResponseEntity.noContent().build();
+	}
+	
+	@DeleteMapping("/{id}")
 	@Transactional
 	public ResponseEntity<Void> delete(@PathVariable Long id, @RequestHeader("Authorization") String header) {
-		checkOwnerService.isOwner(id, header);
-		
+		checkOwnerService.checkOwner(id, header);
 		userService.delete(id);
 		
 		return ResponseEntity.noContent().build();
